+++ packages/generate/_test/page_handler_generator.ts
@@ -1,10 +1,8 @@
 import type { Code, GeneratorOptions, RouteModule } from "@http/generate/types";
 import {
-  asCodePattern,
-  code,
+  asFn,
   importNamed,
-  literalOf,
-  relativeModulePath,
+  staticImport,
 } from "@http/generate/code-builder";
 import { hasBodyFunction } from "$test/generate/page_handler_mapper.ts";
 
@@ -16,41 +14,21 @@ export const handlerMapper = "$test/generate/page_handler_mapper.ts";
  * the [example handler mapper](./page_handler_mapper.ts).
  */
 export function generate(
-  { pattern, module, loaded }: RouteModule,
-  { moduleOutUrl, httpModulePrefix, moduleImports }: GeneratorOptions,
+  { module, loaded }: RouteModule,
+  {}: GeneratorOptions,
   i: number,
 ): Code | undefined {
   if (hasBodyFunction(loaded)) {
-    const byPattern = importNamed(
-      `${httpModulePrefix}route/by-pattern`,
-      "byPattern",
-    );
 
-    const pageHandler = importNamed(
+    const pageHandler = asFn(staticImport(importNamed(
       `$test/generate/page_handler.ts`,
       "pageHandler",
-    );
+    )));
 
-    switch (moduleImports) {
-      case "dynamic": {
-        const lazy = importNamed(`${httpModulePrefix}route/lazy`, "lazy");
-
-        return code`${byPattern}(${
-          asCodePattern(pattern)
-        }, ${lazy}(async () => ${pageHandler}((await import(${
-          literalOf(relativeModulePath(module, moduleOutUrl))
-        })).body)))`;
-      }
-
-      case "static": {
-        const routeModule = importNamed(
-          relativeModulePath(module, moduleOutUrl),
-          `page_body_${i}`,
-          "body"
-        );
-
-        return code`${byPattern}(${asCodePattern(pattern)}, ${pageHandler}(${routeModule}))`;
-      }
-    }
+    return pageHandler(importNamed(
+      module,
+      "body",
+      `page_body_${i}`,
+    ));
   }
 }