+++ packages/generate/_test/routes_default_opts.ts
@@ -5,8 +5,4 @@ import { byPattern } from "@http/route/by-pattern";
 import { cascade } from "@http/route/cascade";
 import { lazy } from "@http/route/lazy";
 
-export default cascade(
-  byPattern("/user/:name{/}?", lazy(async () => byMethod(await import("./routes/user/:name/index.ts")))),
-  byPattern("/about", lazy(async () => byMethod(await import("./routes/about.ts")))),
-  byPattern("/", lazy(() => import("./routes/index.ts"))),
-);
+export default cascade(byPattern("/user/:name{/}?", lazy(async () => byMethod(await import("./routes/user/:name/index.ts")))), byPattern("/about", lazy(async () => byMethod(await import("./routes/about.ts")))), byPattern("/", lazy(async () => (await import("./routes/index.ts")).default)));