+++ packages/generate/default_handler_generator.ts
@@ -1,13 +1,6 @@
 import { hasDefaultHandler } from "@http/discovery/default-handler-mapper";
 import type { Code, GeneratorOptions, RouteModule } from "./types.ts";
-import {
-  asCodePattern,
-  code,
-  importDefault,
-  importNamed,
-  literalOf,
-  relativeModulePath,
-} from "./code_builder.ts";
+import { importDefault } from "./code-builder/mod.ts";
 
 export const handlerMapper = "@http/discovery/default-handler-mapper";
 
@@ -16,35 +9,14 @@ export const handlerMapper = "@http/discovery/default-handler-mapper";
  * as the default export.
  */
 export function generate(
-  { pattern, module, loaded }: RouteModule,
-  { moduleOutUrl, httpModulePrefix, moduleImports }: GeneratorOptions,
+  { module, loaded }: RouteModule,
+  {}: GeneratorOptions,
   i: number,
 ): Code | undefined {
   if (hasDefaultHandler(loaded)) {
-    const byPattern = importNamed(
-      `${httpModulePrefix}route/by-pattern`,
-      "byPattern",
+    return importDefault(
+      module,
+      `route_${i}`,
     );
-
-    switch (moduleImports) {
-      case "dynamic": {
-        const lazy = importNamed(`${httpModulePrefix}route/lazy`, "lazy");
-
-        return code`${byPattern}(${
-          asCodePattern(pattern)
-        }, ${lazy}(() => import(${
-          literalOf(relativeModulePath(module, moduleOutUrl))
-        })))`;
-      }
-
-      case "static": {
-        const routeModule = importDefault(
-          relativeModulePath(module, moduleOutUrl),
-          `route_${i}`,
-        );
-
-        return code`${byPattern}(${asCodePattern(pattern)}, ${routeModule})`;
-      }
-    }
   }
 }