+++ packages/generate/methods_handler_generator.ts
@@ -1,13 +1,11 @@
 import { hasMethodHandlers } from "@http/discovery/methods-handler-mapper";
 import type { Code, GeneratorOptions, RouteModule } from "./types.ts";
 import {
-  asCodePattern,
-  code,
+  asFn,
   importAll,
   importNamed,
-  literalOf,
-  relativeModulePath,
-} from "./code_builder.ts";
+  staticImport,
+} from "./code-builder/mod.ts";
 
 export const handlerMapper = "@http/discovery/methods-handler-mapper";
 
@@ -16,42 +14,19 @@ export const handlerMapper = "@http/discovery/methods-handler-mapper";
  * method handling functions.
  */
 export function generate(
-  { pattern, module, loaded }: RouteModule,
-  { moduleOutUrl, httpModulePrefix, moduleImports }: GeneratorOptions,
+  { module, loaded }: RouteModule,
+  { httpModulePrefix }: GeneratorOptions,
   i: number,
 ): Code | undefined {
   if (hasMethodHandlers(loaded)) {
-    const byPattern = importNamed(
-      `${httpModulePrefix}route/by-pattern`,
-      "byPattern",
-    );
-
-    const byMethod = importNamed(
+    const byMethod = asFn(staticImport(importNamed(
       `${httpModulePrefix}route/by-method`,
       "byMethod",
-    );
+    )));
 
-    switch (moduleImports) {
-      case "dynamic": {
-        const lazy = importNamed(`${httpModulePrefix}route/lazy`, "lazy");
-
-        return code`${byPattern}(${
-          asCodePattern(pattern)
-        }, ${lazy}(async () => ${byMethod}(await import(${
-          literalOf(relativeModulePath(module, moduleOutUrl))
-        }))))`;
-      }
-
-      case "static": {
-        const routeModule = importAll(
-          relativeModulePath(module, moduleOutUrl),
-          `route_methods_${i}`,
-        );
-
-        return code`${byPattern}(${
-          asCodePattern(pattern)
-        }, ${byMethod}(${routeModule}))`;
-      }
-    }
+    return byMethod(importAll(
+      module,
+      `route_methods_${i}`,
+    ));
   }
 }